Updated the LeaderboardRecord count field from int to long

+++ src/com/birtek/cashew/database/LeaderboardRecord.java
@@ -1,3 +1,3 @@
 package com.birtek.cashew.database;
 
-public record LeaderboardRecord(int place, String userID, int count) {}
+public record LeaderboardRecord(int place, String userID, long count) {}
